+++ src/aks-preview/azext_aks_preview/_params.py
@@ -126,7 +126,8 @@ from azext_aks_preview._validators import (
     validate_azuremonitorworkspaceresourceid,
     validate_grafanaresourceid,
     validate_ksm_labels,
-    validate_ksm_annotations
+    validate_ksm_annotations,
+    validate_disable_windows_outbound_nat,
 )
 
 # candidates for enumeration
@@ -332,6 +333,7 @@ def load_arguments(self, _):
         c.argument('dns_zone_resource_id')
         c.argument('enable_keda', action='store_true', is_preview=True)
         c.argument('enable_node_restriction', action='store_true', is_preview=True, help="enable node restriction for cluster")
+        c.argument('enable_cilium_dataplane', action='store_true', is_preview=True)
         # nodepool
         c.argument('host_group_id', validator=validate_host_group_id, is_preview=True)
         c.argument('crg_id', validator=validate_crg_id, is_preview=True)
@@ -504,6 +506,7 @@ def load_arguments(self, _):
         c.argument('workload_runtime', arg_type=get_enum_type(workload_runtimes), default=CONST_WORKLOAD_RUNTIME_OCI_CONTAINER)
         c.argument('gpu_instance_profile', arg_type=get_enum_type(gpu_instance_profiles))
         c.argument('enable_custom_ca_trust', action='store_true', validator=validate_enable_custom_ca_trust)
+        c.argument('disable_windows_outbound_nat', action='store_true', validator=validate_disable_windows_outbound_nat)
 
     with self.argument_context('aks nodepool update') as c:
         c.argument('enable_cluster_autoscaler', options_list=[