+++ src/aks-preview/azext_aks_preview/custom.py
@@ -673,6 +673,7 @@ def aks_create(
     enable_keda=False,
     enable_node_restriction=False,
     enable_vpa=False,
+    enable_cilium_dataplane=False,
     # nodepool
     host_group_id=None,
     crg_id=None,
@@ -1170,6 +1171,7 @@ def aks_agentpool_add(
     workload_runtime=None,
     gpu_instance_profile=None,
     enable_custom_ca_trust=False,
+    disable_windows_outbound_nat=False,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()